# openapi-schema.yaml
openapi: 3.0.0
info:
  title: NaviDocs API
  description: Comprehensive boat documentation management API with 16 supporting tables
  version: 1.0.0
  contact:
    name: NaviDocs Team
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: https://api.navidocs.com/v1
    description: Production server
paths:
...
  /organizations/{organizationId}/contacts:
    get:
      summary: List contacts for an organization
      operationId: listContacts
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'
    post:
      summary: Create a new contact
      operationId: createContact
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContactRequest'
      responses:
        '201':
          description: Contact created successfully
  /contacts/{contactId}:
    get:
      summary: Get a specific contact
      operationId: getContact
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
    put:
      summary: Update a contact
      operationId: updateContact
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContactRequest'
      responses:
        '200':
          description: Contact updated successfully
    delete:
      summary: Delete a contact
      operationId: deleteContact
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Contact deleted successfully

  # Expenses Endpoints
...
components:
  schemas:
...
    Contact:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        name:
          type: string
        type:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateContactRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        notes:
          type: string

    # Expense Schema
...